requirements.txt
-e git+https://github.com/rtdean/grequests@19239a34b00b8ac226b21f01b0fb55e869097fb7#egg=grequests-0.3.1
CsvSchema==1.1.1
argparse>=1.2.1,<1.4.0
beautifulsoup4==4.3.2
dnspython==1.12.0
feedparser==5.1.3
greenlet>=0.4.2,<0.5.0
logbook==0.8.1
netaddr==0.7.12
pygeoip>=0.3.1,<0.4.0
requests>=2.3.0,<2.6.0
sortedcontainers==0.9.4
unicodecsv==0.9.4
wsgiref==0.1.2
yapsy==1.10.423
